split classes in separate files * update namings
