+++ src/scss/skin-modern/_skin.scss
@@ -17,6 +17,8 @@
   @import 'components/label';
   @import 'components/settingspanel';
   @import 'components/settingspanelpage';
+  @import 'components/settingspanelpageopenbutton';
+  @import 'components/settingspanelpagebackbutton';
   @import 'components/settingstogglebutton';
   @import 'components/selectbox';
   @import 'components/listbox';
