+++ src/scss/skin-modern/components/subtitlesettings/_subtitlesettingsopenbutton.scss
@@ -2,16 +2,5 @@
 @import '../../mixins';
 
 .#{$prefix}-ui-subtitlesettingsopenbutton {
-  @extend %ui-button;
-
-  background-image: svg('assets/skin-modern/images/settings.svg');
-  padding: .3em;
-
-  &:hover {
-    @include svg-icon-shadow;
-  }
-
-  &.#{$prefix}-on {
-    background-image: svg('assets/skin-modern/images/settingsX.svg');
-  }
+  @extend %ui-settingspanelpageopenbutton;
 }
